+++ comicfn2dict/regex.py
@@ -72,4 +72,4 @@ ISSUE_BEGIN_RE = re_compile(r"^(" + _ISSUE_RE_EXP + r")\b")
 ISSUE_ANYWHERE_RE = re_compile(r"\b(" + _ISSUE_RE_EXP + r")\b")
 
 # LONG STRINGS
-REMAINING_GROUP_RE = re_compile(r"^[\w].*[^\)]")
+REMAINING_GROUP_RE = re_compile(r"^[^\()].*[^\)]")